### 2017-01-04 #152 @whedon
[PRE REVIEW]: AncesTrim - a tool for trimming complex pedigrees
**Submitting author:** @JNisk (<a href="http://orcid.org/0000-0002-2871-6890">Julia Niskanen</a>)
**Repository:** <a href="https://github.com/JNisk/AncesTrim" target ="_blank">https://github.com/JNisk/AncesTrim</a>
**Version:** v1.0

Thanks for submitting your paper to JOSS JNisk. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
