### 2017-01-04 aeda0ce
Update paper.md
