+++ paper/paper.md
@@ -11,6 +11,12 @@ authors:
  - name: Hannes Lohi
    orcid: 0000-0003-1087-5532
    affiliation: 1, 2, 3
+ - name: Elina Salmela
+   orcid: 
+   affiliation: 
+ - name: Hannes Lohi
+   orcid: 0000-0003-1087-5532
+   affiliation: 1, 2, 3
 affiliations:
  - name: Department of Veterinary Biosciences, University of Helsinki, Helsinki, Finland
    index: 1
