### 2017-01-04 da24739
Update paper.md
