+++ paper/paper.md
@@ -13,7 +13,7 @@ authors:
    affiliation: 1, 2, 3
  - name: Elina Salmela
    orcid: 
-   affiliation: 
+   affiliation: 1, 2, 3, 4
  - name: Hannes Lohi
    orcid: 0000-0003-1087-5532
    affiliation: 1, 2, 3
@@ -24,13 +24,15 @@ affiliations:
    index: 2
  - name: The Folkhälsan Institute of Genetics, Helsinki, Finland
    index: 3
+ - name: Department of Biosciences, University of Helsinki, Helsinki, Finland
+   index: 4
 date: 3 January 2017
 bibliography: paper.bib
 ---
 
 # Summary
 
-AncesTrim (Niskanen & Lohi 2017) is a Python tool for trimming complex pedigrees. Some pedigree data
+AncesTrim (Niskanen et al. 2017) is a Python tool for trimming complex pedigrees. Some pedigree data
 can have an inherently complicated structure with a high level of inbreeding, which makes
 visualization of such data difficult. AncesTrim reduces the complexity of pedigrees while
 preserving critical relatedness structures, allowing the pedigree to be displayed in a more
@@ -38,4 +40,4 @@ efficient way.
 
 # References
 
-Niskanen, J. and Lohi, H. (2017). AncesTrim - a tool for trimming complex pedigrees. https://github.com/JNisk/AncesTrim.
+Niskanen, J., Salmela, E., Lohi, H. (2017). AncesTrim - a tool for trimming complex pedigrees. https://github.com/JNisk/AncesTrim.
