### 2017-01-04 79bb737
Update paper.md
