### 2017-01-04 10e936f
Update paper.bib
