+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @online{AncesTrim,
-  author = {Julia Niskanen and Hannes Lohi},
+  author = {Niskanen, J., Salmela, E., Lohi, H.},
   title = {AncesTrim - a tool for trimming complex pedigrees},
   year = 2017,
   url = {https://github.com/JNisk/AncesTrim},
