### 2017-01-04 8f8853c
Update paper.md
