### 2017-01-06 #152 @mgymrek
I can take this one.
