### 2017-02-07 #179 @whedon
[REVIEW]: AncesTrim - a tool for trimming complex pedigrees
**Submitting author:** @JNisk (<a href="http://orcid.org/0000-0002-2871-6890">Julia Niskanen</a>)
**Repository:** <a href="https://github.com/JNisk/AncesTrim" target ="_blank">https://github.com/JNisk/AncesTrim</a>
**Version:** v1.0

- [x] **Authorship:** Has the submitting author (JNisk) made major contributions to the software?
