### 2017-02-18 #179 @jyuan1322
Overall I found AncesTrim easy to install and run, and would only suggest minor changes to code documentation and error handling.

Functionality:
I tested AncesTrim using Python 2.7 on both an Ubuntu 14.04 and a MacOSX 10.11 machine and found no issues. The code functions as expected and performance is fast.

Documentation:
- The statement of need could be expanded to make the utility of the tool more apparent. The initial paragraph of the README.md is brief and lacking in specifics about what exactly the tool does. Simply moving some of the information under "Pedigree Trimming Principles" into this section might help, but it would be great if the authors could include a visualization of a pedigree before and after pruning with AncesTrim to effectively demonstrate functionality.

Code:
- There should be more extensive error checking when parsing input files. For example, submitting an empty register file will throw a "list index out of range" error.

- In general, the code in ancestrim.py is difficult to follow. While steps are broken into distinct blocks of code, I think these should be wrapped into functions, with clear descriptions of the inputs and outputs of each step.

- The command line options can be improved in a few ways. The authors have written a parser from scratch, but using an existing module like argparse would improve understandability and maintenance of the code and reduce the possibility of error. Also, I had to look up command line options by looking in the code - there should be a way to print a manual page either by running the script without arguments or by supplying a "--help" argument. Lastly, the "--folder" argument should only pertain to an output directory, not one for both input and output files. Since the input filenames need to be supplied anyway, this is a bit redundant and prevents tab-completion of filenames.

- The raw input file appears to follow a custom specification, so users will need to preprocess their data into the expected format. The usability of the tool could be improved by adding a parser for existing pedigree data formats, such as GEDCOM, as this is already generated as an output file. (Although I wouldn't consider this a necessary addition.)

- The code and column headers of the input/output files make multiple references to dogs, when presumably AncesTrim can also operate on pedigrees of humans or other animals.

Minor issues:
- Typo in README.md under Usage: "(the raw rile)"
