### 2017-02-20 #179 @mgymrek
Thanks @jyuan1322 for the review!

@JNisk: it would be great if you could respond to these points. Most importantly:

* making the suggested improvements to the README
* including a "--help" option to list all possible command-line arguments
