### 2017-02-20 #179 @JNisk
Dear @jyuan1322 and @mgymrek,

thank you very much for your constructive criticism. I will now continue working with the script to address these issues.

Best regards,
JNisk
