### 2017-02-23 a35edf2
Update ancestrim.py
