+++ ancestrim_module.py
@@ -36,7 +36,24 @@ def handle_unicode(name):
 
 def parse_commandline():
 	"""Parse parameters from command line with syntax --option [filename] and return a dict object."""
-	def_args = {"--out":"output file name", "--folder":"subfolder"}
+	if "--help" in sys.argv:
+		help_dict = {}
+		help_dict["--raw"] = "File name of the file that contains pedigree information."
+		help_dict["--register"] = "File name of the file that contains the list of individuals for whom the pedigree will be constructed."
+		help_dict["--gen"] = "A numeric parameter that determines the maximum number of generations that will be included in the pedigree."
+		help_dict["--out"] = "File name for the output files that will be generated, without a filename extension."
+		help_dict["--outfolder"] = "Folder name for output files."
+		print("AncesTrim v1.1\n..............\n")
+		print("The following command line parameters are required to run the script:")
+		for i in help_dict:
+			print(" " + i + "  " + help_dict[i])
+		help_opt_dict = {}
+		help_opt_dict["--infolder"] = "Folder name for input files."
+		print("\nThe following optional parameter can also be used:")
+		for i in help_opt_dict:
+			print(" " + i + "  " + help_opt_dict[i])
+		exit(0)
+	def_args = {"--out":"output file name", "--outfolder":"subfolder"}
 	if "ancestrim" in sys.argv[0]:
 		def_args["--raw"] = "raw file"
 		def_args["--register"] = "register file"
@@ -49,18 +66,22 @@ def parse_commandline():
 		if d not in arg_keys:
 			print("missing option " + str(d))
 			print("please specify " + str(def_args[d]) + " with " + str(d))
+			print("\nfor help with command line parameters, use the parameter --help")
 			func_exit()
 	arg_options = {}
 	for i in arguments:
 		#assign options as keys and file names as values in dict arg_options
 		try:
-			for j in ["--out","--folder","--gen","--raw","--register"]:
+			for j in ["--out","--outfolder","--gen","--raw","--register"]:
 				if str(i[0]) == j:
 					arg_options[j] = str(i[1])
 		except IndexError:
 			print("error in options, please check command line")
 			func_exit()
-        forbidden_characters = ["\\","/",":","*","?",'"',"<",">","|","#"]
+		# optionally, include argument --infolder
+		if str(i[0]) == "--infolder":
+			arg_options[str(i[0])] = str(i[1])
+	forbidden_characters = ["\\","/",":","*","?",'"',"<",">","|","#"]
 	for i in forbidden_characters:
 		if i in arg_options["--out"]:
 			print('invalid outfile name: file names cannot contain \/:*?"<>|#')
@@ -76,12 +97,19 @@ def open_rawfile(rawfile,folder):
 	o = open((str(folder) + str(rawfile)),"r")
 	header = o.readline().strip().split("\t")
 	dog_data = {}
+	line_count = 1
 	for line in o:
+		line_count += 1
 		line = line.rstrip("\n").split("\t")
 		line = [i.rstrip() for i in line]
 		tmp = {}
 		for i in header:
-			tmp[i] = handle_unicode(line[header.index(i)])
+			try:
+				tmp[i] = handle_unicode(line[header.index(i)])
+			except IndexError:
+				print("error in reading raw file on line " + str(line_count))
+				print("please check raw file")
+				exit(0)
 		dog_data[tmp["REGISTER_NUMBER"]] = tmp
 	return dog_data
 
@@ -90,15 +118,21 @@ def open_rawfile(rawfile,folder):
 def open_registerfile(regfile,folder):
[…]
