### 2017-02-23 #179 @JNisk
To respond to the points made by @jyuan1322:

>* The statement of need could be expanded to make the utility of the tool more apparent. The initial paragraph of the README.md is brief and lacking in specifics about what exactly the tool does. Simply moving some of the information under "Pedigree Trimming Principles" into this section might help, but it would be great if the authors could include a visualization of a pedigree before and after pruning with AncesTrim to effectively demonstrate functionality.

The first section of README has been modified to be more descriptive. In addition, a "before and after" image has been added for demonstration. I had thought of an image before and hesitated, but your suggestion made it apparent that this improves the README quite a bit. Also, the "Pedigree trimming principles" section now includes an example image.

> * There should be more extensive error checking when parsing input files. For example, submitting an empty register file will throw a "list index out of range" error.

Error checking has been improved. Both raw file and register file are checked and an error message is displayed if either file is empty. Empty lines in register file are skipped, and lines in raw file are checked for missing columns. Also, an error message is displayed if an individual exists in the register file but not in the raw file.

> * In general, the code in ancestrim.py is difficult to follow. While steps are broken into distinct blocks of code, I think these should be wrapped into functions, with clear descriptions of the inputs and outputs of each step.

I agree with this. From my point of view, the complexity stems from the numerous pairwise comparisons that are made while searching for relatedness and when trimming the relationship paths. I understand that wrapping these into functions could make the script easier to read. However, I imagine this tool being used as a whole, which lessens the need to make particular steps available as functions. But in general this is a very valuable piece of advice and I aim to be more function-orientated in the future.

> * The command line options can be improved in a few ways. The authors have written a parser from scratch, but using an existing module like argparse would improve understandability and maintenance of the code and reduce the possibility of error. Also, I had to look up command line options by looking in the code - there should be a way to print a manual page either by running the script without arguments or by supplying a "--help" argument. Lastly, the "--folder" argument should only pertain to an output directory, not one for both input and output files. Since the input filenames need to be supplied anyway, this is a bit redundant and prevents tab-completion of filenames.

The command line options have been improved. A “--help” argument has been added, and it will display information about all the available parameters. Also, if a parameter is missing, messages about both the missing argument and the “--help” option will be displayed. This will also happen if no arguments are provided.

The “--folder” argument has been modified. The script now only requires an “--outfolder” parameter that designates the destination of the output files. An optional parameter “--infolder” can be used in case the user wants to specify the input file folder instead of tab-completion.

This custom command line parser was constructed as a practice, but I know realize that it would indeed have been better to use the argparse module. I am not keen on completely changing the parser at this point, but I will definitely change my approach for future projects. It is a very good point.

> * The raw input file appears to follow a custom specification, so users will need to preprocess their data into the expected format. The usability of the tool could be improved by adding a parser for existing pedigree data formats, such as GEDCOM, as this is already generated as an output file. (Although I wouldn't consider this a necessary addition.)

This is an excellent suggestion, and I had not thought of this before. The lack of input file parsing is partly because this script is part of a custom pedigree manipulation pipeline, but I understand that it would be a valuable addition to this tool.

> * The code and column headers of the input/output files make multiple references to dogs, when presumably AncesTrim can also operate on pedigrees of humans or other animals.

This is something that I considered, and it is true that AncesTrim can process any pedigree data regardless of species. I decided against changing the column names (and, additionally, the variable names in the script itself) as I thought that it is a minor issue that does not affect the performance of the script. I agree that the columns seem a bit silly, though, if one is working with something else than dog data.

> * Typo in README.md under Usage: "(the raw rile)"

The typo has been fixed.

In addition, I made some changes in the files to reflect the version change from 1.0 to 1.1. Changelog has been updated appropriately.

I want to sincerely thank @jyuan1322 for your hard work and great comments!
Best regards,
JNisk
