### 2017-02-27 3e70e4b
Update paper.md
