### 2017-03-07 #179 @mgymrek
@arfon I think we can accept this.

@JNisk I believe next is to make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive.
