### 2017-03-08 #179 @JNisk
I have now archived version 1.1, and the DOI is **10.5281/zenodo.375807**. I also updated the codemeta.json with the DOI.

Best regards,
JNisk
